fix(lint): silence reserved-id on _write retarget; cast int16_t arithmetic-shift result
clang-tidy --warnings-as-errors caught two issues in PR5:
 - cdc_print.c: `_write` is a reserved identifier (cert-dcl37-c). The name
   is mandated by picolibc's retarget contract, so suppress locally with a
   targeted NOLINT rather than renaming.
 - ina238.c: `(int16_t)raw >> 4` integer-promotes to `int` before the shift,
   then narrows back when assigned to `int16_t` (bugprone-narrowing-conversions).
   Cast the shift result explicitly to int16_t to keep the intent.

+++ Software/STMCubeV2/BatteryMonitorV2/Core/Src/ina238.c
@@ -239,7 +239,7 @@ ina238_status_t ina238_read_die_temp_mc(ina238_t *dev, int32_t *milli_celsius)
     /* Top 12 bits are signed °C, bottom 4 reserved. Arithmetic shift on a
      * signed value sign-extends; the compiler we use (gcc/clang) defines
      * this consistently. */
-    int16_t top12 = (int16_t)raw >> 4;
+    int16_t top12 = (int16_t)((int16_t)raw >> 4);
     *milli_celsius = (int32_t)top12 * (int32_t)INA238_DIETEMP_LSB_mC;
     return INA238_OK;
 }
